debug build

+++ fetch_cord/out.py
@@ -133,8 +133,11 @@ def neofetch(loop):
 
     elif not neofetchwin:
         filepath = "/tmp/out.txt"
-        with open(filepath, 'w') as f:
-            print(baseinfo, file=f)
+        if args.config:
+            filepath = args.config
+        if not args.config:
+            with open(filepath, 'w') as f:
+                print(baseinfo, file=f)
         with open(filepath, "rt") as f:
             for line in f:
                 if line.find(cpu) != -1:
@@ -187,7 +190,7 @@ def neofetch(loop):
                     batteryline.append(line.rstrip('\n'))
 
     try:
-        if os.path.isfile(filepath):
+        if os.path.isfile(filepath) and not args.config:
             os.remove(filepath)
     except FileNotFoundError:
         pass
